Removed comments, refactored RobotRegistry, broke project into multiple files
