+++ robot_name/errors.rb
@@ -1,0 +1,3 @@
+class RobotError < RuntimeError; end
+class NameCollisionError < RobotError; end
+class NameValidationError < RobotError; end
